+++ src/renderer/i18n/index.ts
@@ -2,10 +2,10 @@ import i18n from 'i18next';
 import LanguageDetector from 'i18next-browser-languagedetector';
 import { initReactI18next } from 'react-i18next';
 
+import de from './de.json';
 import en from './en.json';
 import fr from './fr.json';
 import lt from './lt.json';
-import de from './de.json';
 
 i18n
   .use(LanguageDetector)
@@ -15,7 +15,7 @@ i18n
       en: { translation: en },
       lt: { translation: lt },
       fr: { translation: fr },
-      fr: { translation: de }
+      de: { translation: de }
     },
     lng: 'en',
     fallbackLng: 'en',